rtu transport: flush RX buffer on init
Some serial drivers seem to queue received data while the port is
closed and are happy to deliver it as reply to our first request,
making it either fail with some error (best case) or succeed with
outdated data (worst case, very low probability of happening).

This fixes it by adding a call to discard() on the RTU transport's
underlying link (socket or serial line) on transport initialization
to flush the RX buffer before sending out our first query.